requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
pymupdf
Pillow
pytesseract
qdrant-client
sentence-transformers
